26 feb

+++ src/main/java/com/example/demo/Players/PlayerRepository.java
@@ -5,11 +5,12 @@ import org.springframework.data.repository.CrudRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface PlayerRepository extends CrudRepository<Player, Integer> {
+public interface PlayerRepository extends CrudRepository< Player, Integer> {
 
     List<Player> findBylastName (String lastName);
     List<Player> findByteamNumber (long teamNumber);
-    List<Player> findByid (long id);
+    List <Player> findByid (Integer id);
+   //Optional <Player> findByid (long id);
